sync up make dist from trunk to pvfs-2-7-branch
SVN revision 6832

+++ maint/make-dist.sh
@@ -19,19 +19,20 @@ if test -z $2; then
     exit 1
 fi
 
+BUILDDIR=$(pwd)
 SRCDIR="$1"
 PVFS2_VERSION="$2"
 
 if test "x$SRCDIR" = "x."; then
-    SRCDIR=`pwd`
+    SRCIR=`pwd`
     echo "Assuming top-level source directory is $SRCDIR"
 fi
 
 STARTDIR=`pwd`
 TARGETBASE="pvfs-$PVFS2_VERSION"
 TARGETDIR="/tmp/$TARGETBASE"
-TARFILE_NAME="$SRCDIR/pvfs-$PVFS2_VERSION.tar"
-TARBALL_NAME="$SRCDIR/pvfs-$PVFS2_VERSION.tar.gz"
+TARFILE_NAME="$BUILDDIR/pvfs-$PVFS2_VERSION.tar"
+TARBALL_NAME="$BUILDDIR/pvfs-$PVFS2_VERSION.tar.gz"
 TAR=`which tar`
 GZIP=`which gzip`
 
@@ -48,18 +49,49 @@ fi
 ##################
 # start processing
 ##################
-make clean
-make docs
-make statecomp
 
+# new directory to construct release in
 if test -d $TARGETDIR; then
     rm -rf $TARGETDIR
 fi
 mkdir $TARGETDIR
+if [ $? -ne 0 ]
+then
+    exit 1
+fi
 
+# copy source over
 cp -f --no-dereference -R $SRCDIR/* $TARGETDIR
 
 cd $TARGETDIR
+
+# dump some special options into the top level module.mk.in
+echo "DIST_RELEASE = 1" >> module.mk.in
+
+# configure, build docs and state machine files
+./configure
+if [ $? -ne 0 ]
+then
+    exit 1
+fi
+make docs
+if [ $? -ne 0 ]
+then
+    exit 1
+fi
+make statecompgen
+if [ $? -ne 0 ]
+then
+    exit 1
+fi
+
+# clean out extra files (distclean will not remove .sm or doc files if
+# DIST_RELEASE is set)
+make distclean
+if [ $? -ne 0 ]
+then
+    exit 1
+fi
 
 # clean out cvs directories and other cruft (if any)
 for f in `find . | grep CVS`; do rm -rf $f; done
@@ -72,11 +104,8 @@ for f in `find . -name module.mk`; do rm -rf $f; done
 for f in `find . -name "*.log"`; do rm -rf $f; done
 for f in `find . -name "*.toc"`; do rm -rf $f; done
 for f in `find . -name "*.aux"`; do rm -rf $f; done
-rm Makefile pvfs2-config.h PVFS2-GLOBAL-TODO.txt
-rm src/common/statecomp/statecomp
-
-# dump some special options into the top level module.mk.in
-echo "DIST_RELEASE = 1" >> module.mk.in
+rm -f Makefile pvfs2-config.h PVFS2-GLOBAL-TODO.txt
+rm -f src/common/statecomp/statecomp
 
 # make sure the cleaned up directory exists
 cd /tmp
